requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.22.0
celery==5.3.1
redis==4.6.0
psycopg[binary,pool]==3.2.9
python-multipart==0.0.9
pdfplumber==0.11.4
prometheus-client==0.22.1
opentelemetry-api==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-instrumentation-fastapi==0.58b0
opentelemetry-instrumentation-celery==0.58b0
opentelemetry-instrumentation-psycopg==0.58b0
opentelemetry-instrumentation-redis==0.58b0
PyYAML==6.0.2
